Break out common test baseline
