+++ tests/codegen/stream/test_stream.py
@@ -1,38 +1,22 @@
 import importlib
-from pathlib import Path
-from typing import AsyncIterable, TextIO
+from typing import AsyncIterable
 
 import pytest
 from pytest_snapshot.plugin import Snapshot
 
 from replit_river.client import Client
-from replit_river.codegen.client import schema_to_river_client_codegen
-from tests.codegen.snapshot.test_enum import UnclosableStringIO
+from tests.codegen.snapshot.codegen_snapshot_fixtures import validate_codegen
 from tests.common_handlers import basic_stream
 
 
 @pytest.mark.parametrize("handlers", [{**basic_stream}])
 async def test_basic_stream(snapshot: Snapshot, client: Client) -> None:
-    snapshot.snapshot_dir = "tests/codegen/snapshot/snapshots"
-    files: dict[Path, UnclosableStringIO] = {}
-
-    def file_opener(path: Path) -> TextIO:
-        buffer = UnclosableStringIO()
-        assert path not in files, "Codegen attempted to write to the same file twice!"
-        files[path] = buffer
-        return buffer
-
-    schema_to_river_client_codegen(
+    validate_codegen(
+        snapshot=snapshot,
         read_schema=lambda: open("tests/codegen/stream/schema.json"),
         target_path="test_basic_stream",
         client_name="StreamClient",
-        file_opener=file_opener,
-        typed_dict_inputs=True,
     )
-
-    for path, file in files.items():
-        file.seek(0)
-        snapshot.assert_match(file.read(), Path(snapshot.snapshot_dir, path))
 
     import tests.codegen.snapshot.snapshots.test_basic_stream
